separate GUI Screen controller for Distraction task
