+++ src/cz/cvut/cognitive/override/DistractionScreenGUIController.java
@@ -16,7 +16,7 @@
 *  along with OpenDS. If not, see <http://www.gnu.org/licenses/>.
 */
 
-package eu.opends.niftyGui;
+package cz.cvut.cognitive.override;
 
 
 import cz.cvut.cognitive.load.CognitiveFunction;
@@ -35,6 +35,7 @@ import cz.cvut.cognitive.distractors.DistractionSettings;
 import cz.cvut.cognitive.distractors.WeatherDistraction;
 import de.lessvoid.nifty.elements.Element;
 import eu.opends.main.Simulator;
+import eu.opends.niftyGui.InstructionScreenGUI;
 
 
 /**
